+++ test/release_packer_gcs_test.dart
@@ -1,6 +1,7 @@
 @TestOn('vm')
-import 'package:release_updater/src/release_packer.dart';
-import 'package:release_updater/src/release_packer_gcs.dart';
+import 'package:release_updater/release_packer_gcs.dart';
+//import 'package:pub_semver/pub_semver.dart' ;
+import 'package:release_updater/release_updater.dart';
 import 'package:test/test.dart';
 
 void main() {
@@ -21,7 +22,9 @@ void main() {
       }
     });
 
-    test('ReleasePackerCommandUploadReleaseBundle', () async {
+    test(
+        'ReleasePackerCommandUploadReleaseBundle (credential: service_account)',
+        () async {
       var cmd = ReleasePackerCommand.from({
         "upload_release": {
           "gcs": {
@@ -39,6 +42,10 @@ void main() {
               "auth_provider_x509_cert_url":
                   "https://www.googleapis.com/oauth2/v1/certs",
               "client_x509_cert_url": "%GCS_CERT_URL%"
+            },
+            "parameters": {
+              "directory": "foo-app/releases",
+              "contentType": "application/bundle",
             }
           }
         }
@@ -55,6 +62,72 @@ void main() {
       expect(cmdGCS.project, equals('project-x'));
       expect(cmdGCS.bucket, equals('project-releases'));
       expect(cmdGCS.credential, isA<Map>());
+
+      {
+        var version = SemanticVersioning.parse('1.0.0');
+        var release = Release("foo-app", version);
+
+        var releaseBundle = ReleaseBundleZip(release, files: [
+          ReleaseFile("foo.sh", "Some script"),
+        ]);
+
+        var bundleBytes = await releaseBundle.toBytes();
+
+        expect(
+            await cmdGCS.resolveUploadParameters(releaseBundle: releaseBundle),
+            equals((
+              bodyBytes: bundleBytes,
+              contentType: 'application/bundle',
+              filePath: 'foo-app/releases/foo-app-1.0.0.zip',
+              release: 'foo-app/1.0.0',
+            )));
+      }
+    });
+
+    test(
+        'ReleasePackerCommandUploadReleaseBundle (credential: metadata.server)',
+        () async {
+      var cmd = ReleasePackerCommand.from({
+        "upload_release": {
+          "gcs": {
+            "project": "project-x",
+            "bucket": "project-releases",
+            "credential": "metadata.server",
+          }
+        }
+      });
+
+      expect(cmd, isA<ReleasePackerCommandUploadReleaseBundle>());
+
+      var cmdBundle = cmd as ReleasePackerCommandUploadReleaseBundle;
+
+      expect(cmdBundle.uploadCommand, isA<ReleasePackerCommandGCS>());
+
+      var cmdGCS = cmdBundle.uploadCommand as ReleasePackerCommandGCS;
+
+      expect(cmdGCS.project, equals('project-x'));
+      expect(cmdGCS.bucket, equals('project-releases'));
+      expect(cmdGCS.credential, isA<String>());
+
+      {
+        var version = SemanticVersioning.parse('1.0.1');
+        var release = Release("foo-app", version);
+
+        var releaseBundle = ReleaseBundleZip(release, files: [
+          ReleaseFile("foo.sh", "Some script"),
+        ]);
+
+        var bundleBytes = await releaseBundle.toBytes();
+
+        expect(
+            await cmdGCS.resolveUploadParameters(releaseBundle: releaseBundle),
+            equals((
+              bodyBytes: bundleBytes,
+              contentType: 'application/zip',
+              filePath: 'foo-app-1.0.1.zip',
+              release: 'foo-app/1.0.1',
+            )));
+      }
     });
   });
 }