+++ lib/src/release_updater_base.dart
@@ -26,7 +26,7 @@ typedef OnRelease = void Function(Release release);
 /// A [Release] updater from [releaseProvider] to [storage].
 class ReleaseUpdater implements Copiable<ReleaseUpdater>, Spawnable {
   // ignore: constant_identifier_names
-  static const String VERSION = '1.1.6';
+  static const String VERSION = '1.1.7';
 
   /// The [Release] storage.
   final ReleaseStorage storage;